UPDATE ::: Fixing logic for converting an RFrame back to an H2Frame using the default method
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@31671 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 015979feca4c4d9230f7f5ff5b8d1c88918871f7

+++ src/prerna/ds/h2/H2Frame.java
@@ -189,9 +189,11 @@ public class H2Frame extends AbstractTableDataFrame {
 		// should delete the folder from the server
 		if(previousPhysicalSchema != null) {
 			File file = new File(previousPhysicalSchema);
-			String folder = file.getParent();
-			logger.info("DELETING ON-DISK SCHEMA AT FOLDER PATH = " + folder);
-			ICache.deleteFolder(folder);
+			if(file.exists()) {
+				String folder = file.getParent();
+				logger.info("DELETING ON-DISK SCHEMA AT FOLDER PATH = " + folder);
+				ICache.deleteFolder(folder);
+			}
 		}
 	}
